Change owner of sample configuration if create files flags is set.

+++ modules/webservice/server/src/schedule_object.cc
@@ -44,11 +44,12 @@ static void _update_host_schedule_info(host const* hst) {
   if (hst->next_check > scheduling_info.last_host_check)
     scheduling_info.last_host_check = hst->next_check;
 
-  scheduling_info.host_check_interval_total += hst->check_interval * config->get_interval_length();
-  scheduling_info.average_services_per_host =
-    (double)scheduling_info.total_services / (double)scheduling_info.total_hosts;
-  scheduling_info.average_scheduled_services_per_host =
-    (double)scheduling_info.total_scheduled_services / (double)scheduling_info.total_hosts;
+  scheduling_info.host_check_interval_total
+    += (unsigned long)(hst->check_interval * (double)config->get_interval_length());
+  scheduling_info.average_services_per_host
+    = (double)scheduling_info.total_services / (double)scheduling_info.total_hosts;
+  scheduling_info.average_scheduled_services_per_host
+    = (double)scheduling_info.total_scheduled_services / (double)scheduling_info.total_hosts;
 
   scheduling_info.max_host_check_spread = config->get_max_host_check_spread();
 
@@ -56,19 +57,20 @@ static void _update_host_schedule_info(host const* hst) {
   if (config->get_host_inter_check_delay_method() == configuration::state::icd_smart
       && scheduling_info.host_check_interval_total > 0) {
 
-    scheduling_info.average_host_check_interval =
-      (double)scheduling_info.host_check_interval_total
+    scheduling_info.average_host_check_interval
+      = (double)scheduling_info.host_check_interval_total
       / (double)scheduling_info.total_scheduled_hosts;
 
-    scheduling_info.average_host_inter_check_delay =
-      (double)scheduling_info.average_host_check_interval
+    scheduling_info.average_host_inter_check_delay
+      = (double)scheduling_info.average_host_check_interval
       / (double)scheduling_info.total_scheduled_hosts;
 
-    scheduling_info.host_inter_check_delay = scheduling_info.average_host_inter_check_delay;
+    scheduling_info.host_inter_check_delay
+      = scheduling_info.average_host_inter_check_delay;
 
     // calculate max inter check delay and see if we should use that instead.
-    double max_inter_check_delay =
-      (double)(scheduling_info.max_host_check_spread * 60.0)
+    double max_inter_check_delay
+      = (double)(scheduling_info.max_host_check_spread * 60.0)
       / (double)scheduling_info.total_scheduled_hosts;
     if (scheduling_info.host_inter_check_delay > max_inter_check_delay)
       scheduling_info.host_inter_check_delay = max_inter_check_delay;
@@ -96,38 +98,41 @@ static void _update_service_schedule_info(service const* svc) {
   if (svc->next_check > scheduling_info.last_service_check)
     scheduling_info.last_service_check = svc->next_check;
 
-  scheduling_info.service_check_interval_total += svc->check_interval * config->get_interval_length();
-  scheduling_info.average_service_check_interval =
-    (double)scheduling_info.service_check_interval_total
+  scheduling_info.service_check_interval_total
+    += (unsigned long)(svc->check_interval * (double)config->get_interval_length());
+  scheduling_info.average_service_check_interval
+    = (double)scheduling_info.service_check_interval_total
     / (double)scheduling_info.total_scheduled_services;
 
-  scheduling_info.average_service_execution_time =
-    (double)((scheduling_info.average_service_execution_time
+  scheduling_info.average_service_execution_time
+    = (double)((scheduling_info.average_service_execution_time
               * (scheduling_info.total_scheduled_services - 1))
              + svc->execution_time)
     / (double)scheduling_info.total_scheduled_services;
 
 
-  scheduling_info.average_services_per_host =
-    (double)scheduling_info.total_services / (double)scheduling_info.total_hosts;
-  scheduling_info.average_scheduled_services_per_host =
-    (double)scheduling_info.total_scheduled_services / (double)scheduling_info.total_hosts;
+  scheduling_info.average_services_per_host
+    = (double)scheduling_info.total_services / (double)scheduling_info.total_hosts;
+  scheduling_info.average_scheduled_services_per_host
+    = (double)scheduling_info.total_scheduled_services / (double)scheduling_info.total_hosts;
 
-  scheduling_info.max_service_check_spread = config->get_max_service_check_spread();
+  scheduling_info.max_service_check_spread
+    = config->get_max_service_check_spread();
 
   // we determine the service inter-check delay.
   if (config->get_service_inter_check_delay_method() == configuration::state::icd_smart
       && scheduling_info.service_check_interval_total > 0) {
 
-    scheduling_info.average_service_inter_check_delay =
-      (double)scheduling_info.average_service_check_interval
+    scheduling_info.average_service_inter_check_delay
+      = (double)scheduling_info.average_service_check_interval
       / (double)scheduling_info.total_scheduled_services;
 
-    scheduling_info.service_inter_check_delay = scheduling_info.average_service_inter_check_delay;
+    scheduling_info.service_inter_check_delay
+      = scheduling_info.average_service_inter_check_delay;
 
     // calculate max inter check delay and see if we should use that instead.
-    double max_inter_check_delay =
-      (double)(scheduling_info.max_service_check_spread * 60.0)
+    double max_inter_check_delay
+      = (double)(scheduling_info.max_service_check_spread * 60.0)
       / (double)scheduling_info.total_scheduled_services;
     if (scheduling_info.service_inter_check_delay > max_inter_check_delay)
       scheduling_info.service_inter_check_delay = max_inter_check_delay;
@@ -143,8 +148,8 @@ static void _update_service_schedule_info(service const* svc) {
 
   // we determine the service interleave factor.
   if (config->get_service_interleave_factor_method() == configuration::state::ilf_smart) {
-    scheduling_info.service_interleave_factor =
-      (int)(ceil(scheduling_info.average_scheduled_services_per_host));
+    scheduling_info.service_interleave_factor
+      = (int)(ceil(scheduling_info.average_scheduled_services_per_host));
 
     logger(dbg_events, most)
       << "Total scheduled service checks: " << scheduling_info.total_scheduled_services << "\n"
